Revert "Merge pull request #4 from TheLazzoro/Guid"
This reverts commit b73d8fe698f7bd09902d60d1d9f3f8d500375f72, reversing
changes made to b661ab208ef6c72ce55a73a6638fb2cd3434aa69.

+++ BetterTriggers/Controllers/ControllerProject.cs
@@ -916,14 +916,14 @@ namespace BetterTriggers.Controllers
 
 
                 // Adjusts local variable ids
-                List<long> blacklistedIds = new List<long>();
+                List<int> blacklistedIds = new List<int>();
                 ControllerTrigger controller = new ControllerTrigger();
                 var varRefs = controller.GetVariableRefsFromTrigger(element);
                 element.trigger.LocalVariables.ForEach(v =>
                 {
                     var lv = (LocalVariable)v;
-                    long oldId = lv.variable.Id;
-                    long newId = Variables.GenerateId(blacklistedIds);
+                    int oldId = lv.variable.Id;
+                    int newId = Variables.GenerateId(blacklistedIds);
                     Variables.AddLocalVariable(lv);
                     lv.variable.Id = newId;
                     blacklistedIds.Add(newId);
