+++ BetterTriggers/Controllers/ControllerTrigger.cs
@@ -81,12 +81,12 @@ namespace BetterTriggers.Controllers
             return generatedName + ".trg";
         }
 
-        public static string GetTriggerName(long triggerId)
+        public static string GetTriggerName(int triggerId)
         {
             return Triggers.GetName(triggerId);
         }
 
-        public static Trigger GetById(long id)
+        public static Trigger GetById(int id)
         {
             return Triggers.FindById(id).trigger;
         }
