+++ BetterTriggers/Models/EditorData/ExplorerElementFolder.cs
@@ -98,7 +98,7 @@ namespace BetterTriggers.Models.EditorData
             }
         }
 
-        public long GetId()
+        public int GetId()
         {
             throw new NotImplementedException();
         }
