+++ BetterTriggers/Models/EditorData/ExplorerElementVariable.cs
@@ -109,7 +109,7 @@ namespace BetterTriggers.Models.EditorData
             }
         }
 
-        public long GetId()
+        public int GetId()
         {
             return variable.Id;
         }
